added order dashboard

+++ Galaxy/src/app/components/admin/brand/main-admin-brand/main-admin-brand.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { UnitService } from '../../../../services/unit.service';
+import { ConfirmMessageComponent } from '../../../shared-componentes/confirm-message/confirm-message.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-main-admin-brand',
@@ -7,7 +10,9 @@ import { UnitService } from '../../../../services/unit.service';
   styleUrls: ['./main-admin-brand.component.css']
 })
 export class MainAdminBrandComponent implements OnInit {
-  constructor(private unit: UnitService) { }
+  constructor(private unit: UnitService, private route: ActivatedRoute,
+    private router: Router,
+    public dialog: MatDialog) { }
   brands: any[] = [];
   isloading: boolean = true;
 
@@ -15,19 +20,35 @@ export class MainAdminBrandComponent implements OnInit {
     this.GetBrands();
   }
 
-  DeleteBrand(brand: any) {
-    this.unit.brand.DeleteBrand(brand.id).subscribe(() => {
-      brand.isDeleted = true;
-    }, error => {
-      console.error('Error deleting brand', error);
+  deleteBrand(brand: any): void {
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to delete ${brand.name}?` }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.unit.brand.DeleteBrand(brand.id).subscribe(() => {
+          brand.isDeleted = true;
+        }, error => {
+          console.error('Error deleting brand', error);
+        });
+      }
     });
   }
 
-  RetriveBrand(brand: any) {
-    this.unit.brand.RetriveBrand(brand.id).subscribe(() => {
-      brand.isDeleted = false;
-    }, error => {
-      console.error('Error retrieving brand', error);
+  retrieveBrand(brand: any): void {
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to retrieve ${brand.name}?` }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.unit.brand.RetriveBrand(brand.id).subscribe(() => {
+          brand.isDeleted = false;
+        }, error => {
+          console.error('Error retrieving brand', error);
+        });
+      }
     });
   }
 
@@ -40,4 +61,5 @@ export class MainAdminBrandComponent implements OnInit {
       this.isloading = false;
     });
   }
+
 }
