+++ Galaxy/src/app/components/admin/product/main-admin-product/main-admin-product.component.ts
@@ -5,6 +5,7 @@ import { IProducts } from '../../../../Models/Product/products-model';
 import { IProductQuery, initProductQuery } from '../../../../Models/Product/product-query-model';
 import { MatDialog } from '@angular/material/dialog';
 import { DetailsAdminProductComponent } from '../details-admin-product/details-admin-product.component';
+import { ConfirmMessageComponent } from '../../../shared-componentes/confirm-message/confirm-message.component';
 
 @Component({
   selector: 'app-main-admin-product',
@@ -33,24 +34,36 @@ export class MainAdminProductComponent implements OnInit, OnDestroy {
       console.log('The dialog was closed');
     });
   }
+DeleteProduct(product: any): void {
+  const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+    data: { message: `Are you sure you want to delete ${product.name}?` }
+  });
 
-  DeleteProduct(product: any): void {
-    this.unit.product.DeleteProduct(product.id).subscribe(() => {
-      product.isDeleted = true;
-      this.GetProducts();
-        }, (error: any) => {
+  dialogRef.afterClosed().subscribe(result => {
+    if (result) {
+      this.unit.product.DeleteProduct(product.id).subscribe(() => {
+        product.isDeleted = true;
+      }, error => {
         console.error('Error deleting product', error);
       });
-  }
-
+    }
+  });
+}
   RetrieveProduct(product: any): void {
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to retrieve ${product.name}?` }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
     this.unit.product.RetrieveProduct(product.id).subscribe(() => {
         product.isDeleted = false;
       }, (error: any) => {
       console.error('Error retrieving product', error);
     });
   }
-
+});
+}
   GetProducts(): void {
     this.productsSubscription = this.unit.products.GetProducts().subscribe((response: any) => {
       this.products = response.products.sort((a: IProducts, b: IProducts) => b.id - a.id);
